Ring Of Wealth rework

+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/mobs/Mob.java
@@ -52,6 +52,7 @@ import com.shatteredpixel.shatteredpixeldungeon.items.bombs.Bomb;
 import com.shatteredpixel.shatteredpixeldungeon.items.bombs.StormBomb;
 import com.shatteredpixel.shatteredpixeldungeon.items.fishingrods.FishingRod;
 import com.shatteredpixel.shatteredpixeldungeon.items.modules.RewardBoostModule;
+import com.shatteredpixel.shatteredpixeldungeon.items.rings.Ring;
 import com.shatteredpixel.shatteredpixeldungeon.items.rings.RingOfWealth;
 import com.shatteredpixel.shatteredpixeldungeon.items.spells.HandyBarricade;
 import com.shatteredpixel.shatteredpixeldungeon.items.spells.MagicBridge;
@@ -895,7 +896,22 @@ public abstract class Mob extends Char {
 
 	public float lootChance(){
 
-		return this.lootChance;
+		float lootChance = this.lootChance;
+
+		float dropBonus = RingOfWealth.dropChanceMultiplier( Dungeon.hero );
+
+		Talent.BountyHunterTracker bhTracker = Dungeon.hero.buff(Talent.BountyHunterTracker.class);
+		if (bhTracker != null){
+			Preparation prep = Dungeon.hero.buff(Preparation.class);
+			if (prep != null){
+				// 2/4/8/16% per prep level, multiplied by talent points
+				float bhBonus = 0.02f * (float)Math.pow(2, prep.attackLevel()-1);
+				bhBonus *= Dungeon.hero.pointsInTalent(Talent.BOUNTY_HUNTER);
+				dropBonus += bhBonus;
+			}
+		}
+
+		return lootChance * dropBonus;
 	}
 
 	public void rollToDropLoot(){
@@ -1094,6 +1110,17 @@ public abstract class Mob extends Char {
 			int rolls = 15 + DivisionItem.streak_d;
 			ArrayList<Item> bonus = RingOfWealth.tryForBonusDrop(rolls);
 		if (!bonus.isEmpty()) {
+				for (Item b : bonus) Dungeon.level.drop(b, pos).sprite.drop();
+				RingOfWealth.showFlareForBonusDrop(sprite);
+			}
+		}
+
+		if (Ring.getBuffedBonus(Dungeon.hero, RingOfWealth.Wealth.class) > 0) {
+			int rolls = 5;
+			if (properties.contains(Property.BOSS)) rolls = 20;
+			else if (properties.contains(Property.MINIBOSS)) rolls = 10;
+			ArrayList<Item> bonus = RingOfWealth.tryForBonusDrop(rolls);
+			if (bonus != null && !bonus.isEmpty()) {
 				for (Item b : bonus) Dungeon.level.drop(b, pos).sprite.drop();
 				RingOfWealth.showFlareForBonusDrop(sprite);
 			}
